+++ app/src/main/java/com/example/laugh/ui/MainActivity.kt
@@ -1,7 +1,8 @@
-package com.example.laugh
+package com.example.laugh.ui
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.example.laugh.R
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
