+++ app/src/main/java/com/example/laugh/view/LoginView.kt
@@ -1,4 +1,4 @@
-package com.example.laugh.login
+package com.example.laugh.view
 
 interface LoginView {
     fun loginSuccess()
