+++ app/src/main/res/layout/activity_authorization.xml
@@ -12,7 +12,7 @@
     android:paddingHorizontal="16dp"
 
     android:background="@color/midnight_express"
-    tools:context=".login.LoginActivity">
+    tools:context=".ui.LoginActivity">
 
     <ScrollView
         android:id="@+id/auth_scroll_fields"
